A stack walker for Xen domains
This tool takes a Xen domain, then pauses it to access the current stack
from outside, producing a stack walk. This is especially useful as a
profiling (or a simple debugging) tool for unikernels. The output it
produces is designed to be used as input to Brendan Gregg's flamegraph
tool.

This commit is still missing some vital functionality, such as the
periodic pausing/unpausing to create many stack trace samples to produce
meaningful profiling output.