uniprof: Fix occasional 0 addresses in trace
Return address was always read from (frame pointer + word size), which
is correct, except when this traverses a xen page. Since we need to map
the pages over from the guest to dom0, _and_ we are not guaranteed that
those mapped pages are contiguous in dom0's memory, we need to treat
that special case, and do an extra mapping for the return address then.

Fixes gitlab issue #1.