arm-support: don't hold you breath for libxc support
Since libxc has support for memory translation for x86, but not arm,
there is surrently no way for libxc to support tracing on arm. I might
get around to write a libxc version of my page traversal function once I
have a fully working version of it for libxencall... but don't wait with
bated breath.