prevent segfault on certain broken symbol table files
If the symbol contains not a single newline character (which trivially
means it is invalid), read_symbol_table() would not create a symbol
table (what from, after all?), but also not properly report this problem
back. During the actual stack walk, trying to resolve the an address to
a symbol would then lead to a segfault.

Either the tests in read_symbol_table() or in binsearch_find_*() would have
been sufficient for this specific bug, but each of them makes a lot of
sense in general anyway.