Stop profiling if domain has shut down
If a domain has shut down, there is nothing to profile any more.
However, Xen doesn't completely destroy the domain yet, because a
certain someone (i.e., we) have still memory mapped from that domain.
How nice of them to not just throw us out and let us crash trying to
read unaccessible memory!

However, such a domain can still be paused and unpaused (for whatever
reason), uniprof doesn't really notice that the domain is, for all
intents and purposes, dead. This results in uniprof continuing to run,
"profiling" the same final stack over and over again.

This being patently silly, we now check whether the domain has shut
down. We only do this once a second though, to keep the overhead of that
check low.