change signature of get_max_vcpu_id() and get_word_size()
Instead of returning a signed integer that denotes failurs (<0) or
success with the desired value, make the return value a success code,
and return the desired value in a pointer to an unsigned int passed into
the function as an argument.

Cleans up the interfaces and reduces typecasts when interfacing with the
xen toolstack libraries (which expect unsigned ints).