use unsigned int for vcpu variable
Cleans up the interface to the xen toolstack libraries (which expect an
unsigned int).

Note that this produces a compile-time warning calling _UXEN_change_vcpu
from libunwind-xen. This warning is fixed in a newer version of the
library, changing that function's signature to accept an unsigned int
instead of an int.