remove superfluous memory unmap
I can't think of any situation where this would unmap memory that isn't
already unmapped. Each map in xen_translate_foreign_address immediately
does an unmap after a memcpy, and even in case the map fails, the
(potentially not even mapped) memory is unmapped on the way out of the
function.